requirements.txt
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.2.0
streamlit==1.8.1
